Refactor - Combine Condition and CompositeCondition classes to simplify our config
Previously there were two different classes, Condition and Composite condition, and using them in pages-config could get pretty complicated. This change aims to simplify our configuration by merging those classes.

Advantages of this change:
 - Now we don't need to have corresponding levels of single conditions and aggregate conditions
 - No longer need ifCompositeCondition checks

[#177891567]
