+++ src/main/java/org/codeforamerica/shiba/pages/config/ApplicationConfiguration.java
@@ -1,8 +1,7 @@
 package org.codeforamerica.shiba.pages.config;
 
 import lombok.Data;
-import org.codeforamerica.shiba.inputconditions.Condition;
-import org.codeforamerica.shiba.output.CompositeCondition;
+import org.codeforamerica.shiba.output.Condition;
 
 import java.util.List;
 import java.util.Map;
@@ -10,7 +9,6 @@ import java.util.Map;
 @Data
 public class ApplicationConfiguration {
     private List<Condition> conditionDefinitions;
-    private List<CompositeCondition> compositeConditionDefinitions;
     private List<PageConfiguration> pageDefinitions;
     private LandmarkPagesConfiguration landmarkPages;
     private Map<String, PageWorkflowConfiguration> workflow;
