+++ src/main/java/org/codeforamerica/shiba/pages/config/ConditionalValue.java
@@ -1,16 +1,10 @@
 package org.codeforamerica.shiba.pages.config;
 
 import lombok.Data;
-import org.codeforamerica.shiba.inputconditions.Condition;
-import org.codeforamerica.shiba.output.CompositeCondition;
+import org.codeforamerica.shiba.output.Condition;
 
 @Data
 public class ConditionalValue {
-    private CompositeCondition compositeCondition;
     private Condition condition;
     private String value;
-
-    public boolean isCompositeCondition(CompositeCondition compositeCondition) {
-        return compositeCondition != null;
-    }
 }
