+++ src/main/java/org/codeforamerica/shiba/pages/config/FormInput.java
@@ -1,7 +1,7 @@
 package org.codeforamerica.shiba.pages.config;
 
 import lombok.Data;
-import org.codeforamerica.shiba.inputconditions.Condition;
+import org.codeforamerica.shiba.output.Condition;
 
 import java.util.Collections;
 import java.util.List;
