+++ src/main/java/org/codeforamerica/shiba/pages/config/PageGroupConfiguration.java
@@ -1,7 +1,7 @@
 package org.codeforamerica.shiba.pages.config;
 
 import lombok.Data;
-import org.codeforamerica.shiba.inputconditions.Condition;
+import org.codeforamerica.shiba.output.Condition;
 
 import java.util.List;
 import java.util.Map;
