+++ src/main/java/org/codeforamerica/shiba/pages/config/PageWorkflowConfiguration.java
@@ -1,7 +1,7 @@
 package org.codeforamerica.shiba.pages.config;
 
 import lombok.Data;
-import org.codeforamerica.shiba.output.CompositeCondition;
+import org.codeforamerica.shiba.output.Condition;
 import org.codeforamerica.shiba.pages.data.ApplicationData;
 import org.codeforamerica.shiba.pages.data.Subworkflows;
 
@@ -11,7 +11,7 @@ import java.util.List;
 @Data
 public class PageWorkflowConfiguration {
     private List<NextPage> nextPages;
-    private CompositeCondition skipCondition;
+    private Condition skipCondition;
     private List<PageDatasource> datasources = new ArrayList<>();
     private PageConfiguration pageConfiguration;
     private String groupName;
