+++ src/main/java/org/codeforamerica/shiba/pages/data/DatasourcePages.java
@@ -1,10 +1,8 @@
 package org.codeforamerica.shiba.pages.data;
 
-import org.codeforamerica.shiba.inputconditions.Condition;
-import org.codeforamerica.shiba.output.CompositeCondition;
+import org.codeforamerica.shiba.output.Condition;
 
 import java.util.HashMap;
-import java.util.Map;
 import java.util.stream.Stream;
 
 public class DatasourcePages extends HashMap<String, PageData> {
@@ -13,6 +11,14 @@ public class DatasourcePages extends HashMap<String, PageData> {
     }
 
     public Boolean satisfies(Condition condition) {
+        if (condition.getConditions() != null) {
+            Stream<Condition> conditionStream = condition.getConditions().stream();
+            return switch (condition.getLogicalOperator()) {
+                case AND -> conditionStream.allMatch(this::satisfies);
+                case OR -> conditionStream.anyMatch(this::satisfies);
+            };
+        }
+
         PageData pageData = this.get(condition.getPageName());
         if (pageData == null || !pageData.containsKey(condition.getInput())) {
             // This page's skipCondition was satisfied, so the client didn't provide an answer, so this condition can't be satisfied
@@ -20,22 +26,6 @@ public class DatasourcePages extends HashMap<String, PageData> {
         } else {
             return condition.matches(pageData, new PagesData(this));
         }
-    }
-
-    public Boolean satisfies(CompositeCondition compositeCondition) {
-        if (compositeCondition.getCompositeConditions() != null) {
-            Stream<CompositeCondition> conditionStream = compositeCondition.getCompositeConditions().stream();
-            return switch (compositeCondition.getLogicalOperator()) {
-                case AND -> conditionStream.allMatch(this::satisfies);
-                case OR -> conditionStream.anyMatch(this::satisfies);
-            };
-        }
-
-        Stream<Condition> conditionStream = compositeCondition.getConditions().stream();
-        return switch (compositeCondition.getLogicalOperator()) {
-            case AND -> conditionStream.allMatch(this::satisfies);
-            case OR -> conditionStream.anyMatch(this::satisfies);
-        };
     }
 
     public DatasourcePages mergeDatasourcePages(DatasourcePages datasourcePages) {
